Remove logz plugin from the image
We want to decommission logz plugin completely so will
remove the logz plugin. The irony is that this repo is
originally for fluentd with logz. Not switching the base
image at this time simply because there are other set up
we have done here and we want to keep the scope to minimal